Delegate problem formatting to a reserved class, making the Problem class slimmer
